alg/discriminative-jackknife/requirements.txt
matplotlib==3.3.2
notebook==6.1.5
numpy==1.19.2
pandas==1.1.5
scikit-learn==0.23.2
scipy==1.5.2
tensorflow-gpu==1.15.0
Theano==1.0.5
torch==1.4.0